vc: generate unique libvirtRoot based on UUID
In order to be able to run multiple containers side by side, we
need the libvirtRoot for each VM to be different. The obvious
solution is to include the VM's UUID in the path, but we have to
be a bit careful with the length or we will reach the maximum
length allowed for sockets: dropping all dashes doesn't affect
the uniqueness of the path and allows us to remain below the
length limits.